# swagger.yaml
swagger: "2.0"

info:
  version: "0.0.1"
  title: "Battleship"
  description: "A simple game to practice node.js"

schemes:
  - "http"
host: "localhost:3000"

paths:
  /boards/playerA:
    get:
      summary: "Get board A"
      description: "Shows board A"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Board"
    put:
      summary: "Drop the old and reload board A with new elements"
      parameters:
        - name: "update board"
          in: "body"
          required: true
          schema:
            "$ref": "#/definitions/Board"
      responses:
        200:
          description: "Successful operation"
          schema:
            "$ref": "#/definitions/Board"

  /boards/playerB:
    get:
      summary: "Get board B"
      description: "Shows board B"
      responses:
        200:
          description: "Successful operation"
    put:
      summary: "Drop the old and reload board B with new elements"
      parameters:
        - name: "new board"
          in: "body"
          required: true
          schema:
            "$ref": "#/definitions/Board"
      responses:
        200:
          description: "Successful operation"
          schema:
            "$ref": "#/definitions/Board"

  /boards/playerA/shoot:
    post:
      description: "Player shots at cell x:y on board B"
      parameters:
        - name: "coordinates"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              x:
                type: "integer"
              y:
                type: "integer"
      responses:
        200:
          description: "Successful operation returns hit or missed"
          schema:
            type: "object"
            properties:
              hit:
                type: "boolean"

  /boards/playerB/shoot:
    post:
      description: "Player shots at cell x:y on board A"
      parameters:
        - name: "coordinates"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              x:
                type: "integer"
              y:
                type: "integer"
      responses:
        200:
          description: "Successful operation returns hit or missed"
          schema:
            type: "object"
            properties:
              hit:
                type: "boolean"

definitions:
  Board:
    type: "array"
    items:
      type: "array"
      items:
        type: "integer"
        enum:
          - 0
          - 1
          - 2
          - 3